Move to server component and rename Messages component so it doesn't colide with Messages Model
